Rename receipt_with_mmr_root  -> receipt_with_mmr_proof
