+++ api/src/ethereum/mod.rs
@@ -4,7 +4,7 @@ mod mmr_leaf;
 mod only_receipt;
 mod parent_mmr_root;
 mod proof;
-mod receipt_with_mmr_root;
+mod receipt_with_mmr_proof;
 
 pub use self::{
     count::handle as count,
@@ -12,5 +12,5 @@ pub use self::{
     only_receipt::handle as only_receipt,
     parent_mmr_root::handle as parent_mmr_root,
     proof::{handle as proof, ProposalReq},
-    receipt_with_mmr_root::{handle as receipt_with_mmr_root, ReceiptResp},
+    receipt_with_mmr_proof::{handle as receipt_with_mmr_proof, ReceiptResp},
 };
