+++ api/src/ethereum/receipt_with_mmr_proof.rs
